imported suggested_books.csv, added functions for cleaning data, changed price's data type to float

+++ models.py
@@ -2,6 +2,7 @@ from sqlalchemy import (create_engine, Column,
                          Integer, String, Date)
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
+from sqlalchemy.sql.sqltypes import Float
 
 
 engine = create_engine('sqlite:///books.db', echo=False)
@@ -16,7 +17,7 @@ class Book(Base):
     title = Column('Title', String)
     author = Column('Author', String)
     published_date = ('Published', Date)
-    price = Column('Price', Integer)
+    price = Column('Price', Float)
 
     def __repr__(self):
         return f'ID= {self.id},Title= {self.title},Author= {self.author},Published= {self.published_date},Price= {self.price}'
